Extract SignatureProofs into their own module
